Feature/portfolio diagramok (#33)
* Create test charts

* Create dynamic pie chart for spot positions

* Create tabs for positions

* Add charts for all portfolio

* Position chart next to balance

* Create method for getting a close price of a currency by date

* Create PerformanceService

* Create performance chart

* valami

* Create method for initing snapshots

* Replace all double variables with BigDecimal

* Change performance to show gains instead of total value

+++ src/main/java/hu/portfoliotracker/DTO/BalanceDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
@@ -17,14 +18,14 @@ public class BalanceDto {
 
     private List<ClosedPositionDto> closedPositionDtos;
 
-    private Double totalDeposit;
+    private BigDecimal totalDeposit;
 
-    private Double totalUnrealizedGains;
+    private BigDecimal totalUnrealizedGains;
 
-    private Double totalUnrealizedGainsPercent;
+    private BigDecimal totalUnrealizedGainsPercent;
 
-    private Double totalRealizedGains;
+    private BigDecimal totalRealizedGains;
 
-    private Double totalRealizedGainsPercent;
+    private BigDecimal totalRealizedGainsPercent;
 
 }
